GraphsFlows and SymRCM extensions
