+++ Project.toml
@@ -7,23 +7,24 @@ version = "0.4.0"
 AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
 Dictionaries = "85a47980-9c8c-11e8-2b9f-f7ca1fa99fb4"
 Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
-GraphsFlows = "06909019-6f44-4949-96fc-b9d9aaa02889"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 PackageExtensionCompat = "65ce6f38-6b18-4e1d-a461-8949797d7930"
-Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 SimpleTraits = "699a6c99-e7fa-54fc-8d76-47d257e15c1d"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 SplitApplyCombine = "03a91e81-4c3e-53e1-a0a4-9c0c8f19dd66"
 Suppressor = "fd094767-a336-5f1f-9728-57cf17d0bbfb"
-SymRCM = "286e6d88-80af-4590-acc9-0001b223b9bd"
 
 [weakdeps]
+GraphsFlows = "06909019-6f44-4949-96fc-b9d9aaa02889"
 KaHyPar = "2a6221f6-aa48-11e9-3542-2d9e0ef01880"
 Metis = "2679e427-3c69-5b7f-982b-ece356f1e94b"
+SymRCM = "286e6d88-80af-4590-acc9-0001b223b9bd"
 
 [extensions]
+NamedGraphsGraphsFlowsExt = "GraphsFlows"
 NamedGraphsKaHyParExt = "KaHyPar"
 NamedGraphsMetisExt = "Metis"
+NamedGraphsSymRCMExt = "SymRCM"
 
 [compat]
 AbstractTrees = "0.3, 0.4"
@@ -34,7 +35,6 @@ KaHyPar = "0.3.1"
 LinearAlgebra = "1.7"
 Metis = "1.4"
 PackageExtensionCompat = "1"
-Requires = "1.3"
 SimpleTraits = "0.9"
 SparseArrays = "1.7"
 SplitApplyCombine = "1.2.2"
@@ -43,8 +43,10 @@ SymRCM = "0.2.1"
 julia = "1.7"
 
 [extras]
+GraphsFlows = "06909019-6f44-4949-96fc-b9d9aaa02889"
 KaHyPar = "2a6221f6-aa48-11e9-3542-2d9e0ef01880"
 Metis = "2679e427-3c69-5b7f-982b-ece356f1e94b"
+SymRCM = "286e6d88-80af-4590-acc9-0001b223b9bd"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
