+++ src/NamedGraphs.jl
@@ -10,8 +10,8 @@ include("decorate.jl")
 include("shortestpaths.jl")
 include("distance.jl")
 include("distances_and_capacities.jl")
-include("steiner_tree/steiner_tree.jl")
-include("traversals/dfs.jl")
+include("steiner_tree.jl")
+include("dfs.jl")
 include("namedgraph.jl")
 include("lib/NamedGraphGenerators/src/NamedGraphGenerators.jl")
 include("lib/PartitionedGraphs/src/PartitionedGraphs.jl")
